% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_convergence.R
\name{check_convergence}
\alias{check_convergence}
\title{Check Convergence of Membership Matrix}
\usage{
check_convergence(U_old, U_new, epsilon = 1e-06)
}
\arguments{
\item{U_old}{Previous membership matrix.}

\item{U_new}{Updated membership matrix.}

\item{epsilon}{Convergence tolerance.}
}
\value{
Logical value indicating convergence.
}
\description{
Evaluates convergence based on the
Frobenius norm between two membership
matrices.
}
\keyword{internal}
