% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_assignment.R
\name{cluster_assignment}
\alias{cluster_assignment}
\title{Hard Cluster Assignment}
\usage{
cluster_assignment(object)
}
\arguments{
\item{object}{An object of class "fcmTFN".}
}
\value{
A factor indicating cluster labels.
}
\description{
Assigns each observation to the cluster
with the highest membership value.
}
