% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_size.R
\name{cluster_size}
\alias{cluster_size}
\title{Cluster Sizes}
\usage{
cluster_size(object)
}
\arguments{
\item{object}{An object of class "fcmTFN".}
}
\value{
A data.frame containing:
\describe{
  \item{Cluster}{Cluster label.}
  \item{Hard_Size}{Number of observations assigned
  to the cluster.}
  \item{Fuzzy_Size}{Sum of membership degrees.}
}
}
\description{
Computes the size of each cluster using both
hard assignments and fuzzy memberships.
}
\details{
Hard size is obtained by assigning each observation
to the cluster with the highest membership degree.

Fuzzy size is computed as the sum of membership
degrees for each cluster.
}
