% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_prototypes.R
\name{compute_prototypes}
\alias{compute_prototypes}
\title{Compute Cluster Prototypes}
\usage{
compute_prototypes(fuzzy_data, U, m = 2)
}
\arguments{
\item{fuzzy_data}{A list of TFN matrices.}

\item{U}{Membership matrix.}

\item{m}{Numeric. Fuzzifier parameter (default = 2).}
}
\value{
A list of prototype matrices.
}
\description{
Computes fuzzy cluster prototypes (centroids)
using membership degrees and triangular fuzzy numbers.
}
\details{
Each prototype is calculated as a weighted average
of fuzzy observations using membership degrees raised
to the fuzzifier parameter m.
}
\keyword{internal}
