% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_weights.R
\name{compute_weights}
\alias{compute_weights}
\title{Compute Adaptive Weights}
\usage{
compute_weights(fuzzy_data, prototypes, U, m = 2)
}
\arguments{
\item{fuzzy_data}{List of TFN matrices.}

\item{prototypes}{List of prototype matrices.}

\item{U}{Membership matrix.}

\item{m}{Fuzzifier parameter.}
}
\value{
A list with wc and ws.
}
\description{
Computes adaptive weights for modal (wc)
and spread (ws) components based on
fuzzy distances and membership degrees.
}
\keyword{internal}
