% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_xie_beni_index.R
\name{compute_xie_beni_index}
\alias{compute_xie_beni_index}
\title{Compute Xie-Beni Cluster Validity Index}
\usage{
compute_xie_beni_index(fuzzy_data, U, prototypes, wc, ws, m = 2)
}
\arguments{
\item{fuzzy_data}{List of triangular fuzzy numbers
representing the dataset.}

\item{U}{Membership matrix (n x k).}

\item{prototypes}{List of cluster prototypes.}

\item{wc}{Modal weight.}

\item{ws}{Spread weight.}

\item{m}{Fuzzifier parameter.}
}
\value{
Numeric value representing the
Xie-Beni index.
}
\description{
Computes the Xie-Beni (XB) index used to evaluate
the compactness and separation of fuzzy clusters.
}
\details{
The Xie-Beni index is defined as:

\deqn{
XB =
\frac{
\sum_{i=1}^{n}
\sum_{g=1}^{k}
u_{ig}^{m} d^{2}(x_i, h_g)
}{
n \min_{g \ne g'}
d^{2}(h_g, h_{g'})
}
}

Lower values indicate better clustering performance.
}
\keyword{internal}
