% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fkm_f_algorithm.R
\name{fkm_f_algorithm}
\alias{fkm_f_algorithm}
\title{Fuzzy K-Means for TFN Data}
\usage{
fkm_f_algorithm(fuzzy_data, k, m = 2, epsilon = 1e-06, max_iter = 1000)
}
\arguments{
\item{fuzzy_data}{List of TFN matrices.}

\item{k}{Number of clusters.}

\item{m}{Fuzzifier parameter.}

\item{epsilon}{Convergence tolerance.}

\item{max_iter}{Maximum iterations.}
}
\value{
List containing clustering results.
}
\description{
Core iterative algorithm for fuzzy clustering
using triangular fuzzy numbers.
}
\keyword{internal}
