% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/membership.R
\name{membership}
\alias{membership}
\title{Extract Membership Matrix}
\usage{
membership(object)
}
\arguments{
\item{object}{An object of class "fcmTFN".}
}
\value{
A matrix where rows represent observations
and columns represent clusters.
}
\description{
Returns the fuzzy membership matrix obtained
from the Fuzzy C-Means clustering process.
}
