% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dictionary.R
\name{plot_dictionary}
\alias{plot_dictionary}
\title{Plot Fuzzy Dictionary}
\usage{
plot_dictionary(object, show_title = TRUE, ...)
}
\arguments{
\item{object}{An object of class "fcmTFN".}

\item{show_title}{Logical. If TRUE (default),
displays the plot title.}

\item{...}{Additional graphical parameters.}
}
\value{
Invisibly returns NULL.
}
\description{
Plots the triangular fuzzy numbers
defining the Likert-scale dictionary.
}
