% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prototypes.R
\name{plot_prototypes}
\alias{plot_prototypes}
\title{Plot Cluster Prototypes}
\usage{
plot_prototypes(
  object,
  cluster = 1,
  view = c("single", "global"),
  use_var_names = FALSE,
  var_names = NULL,
  show_title = TRUE,
  ...
)
}
\arguments{
\item{object}{An object of class "fcmTFN".}

\item{cluster}{Integer cluster to plot when
view = "single".}

\item{view}{Character string indicating the
visualization mode. Either "single" or "global".}

\item{use_var_names}{Logical.}

\item{var_names}{Optional character vector of variable names.}

\item{show_title}{Logical. If TRUE (default),
displays the plot title.}

\item{...}{Additional graphical parameters.}
}
\value{
Invisibly returns NULL.
}
\description{
Visualizes cluster prototypes as interval plots
using triangular fuzzy numbers (l, c, r).
}
