% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_fcmTFN.R
\name{print_fcmTFN}
\alias{print_fcmTFN}
\title{Print Method for fcmTFN Objects}
\usage{
print_fcmTFN(x, ...)
}
\arguments{
\item{x}{An object of class "fcmTFN".}

\item{...}{Additional arguments (not used).}
}
\value{
The input object (invisibly).
}
\description{
Displays a concise summary of a fitted
fuzzy C-means model using triangular fuzzy numbers.
}
