% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototype_matrix.R
\name{prototype_matrix}
\alias{prototype_matrix}
\title{Prototype Matrix Extraction}
\usage{
prototype_matrix(object, use_var_names = FALSE, var_names = NULL)
}
\arguments{
\item{object}{An object of class "fcmTFN".}

\item{use_var_names}{Logical.}

\item{var_names}{Optional variable names.}
}
\value{
A data.frame containing prototype values.
}
\description{
Returns cluster prototypes as a readable matrix
containing l, c, r values for each variable.
}
