% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_fcmTFN.R
\name{summary.fcmTFN}
\alias{summary.fcmTFN}
\title{Summary for fcmTFN Objects}
\usage{
\method{summary}{fcmTFN}(object, ...)
}
\arguments{
\item{object}{An object of class "fcmTFN".}

\item{...}{Additional arguments (not used).}
}
\value{
Prints a formatted summary of the clustering result.
}
\description{
Displays a summary of the Fuzzy C-Means clustering
results for triangular fuzzy numbers.
}
