% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_membership.R
\name{update_membership}
\alias{update_membership}
\title{Update Membership Matrix}
\usage{
update_membership(fuzzy_data, prototypes, wc, ws, m = 2)
}
\arguments{
\item{fuzzy_data}{List of TFN matrices.}

\item{prototypes}{List of cluster prototypes.}

\item{wc}{Modal weight.}

\item{ws}{Spread weight.}

\item{m}{Fuzzifier parameter.}
}
\value{
Updated membership matrix U.
}
\description{
Updates the fuzzy membership matrix (U)
based on distances between observations
and cluster prototypes.
}
\keyword{internal}
