% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_dictionary.R
\name{plot_dictionary}
\alias{plot_dictionary}
\title{Plot Fuzzy Dictionary}
\usage{
plot_dictionary(object)
}
\arguments{
\item{object}{An object of class "fcmTFN".}
}
\value{
A plot showing triangular membership functions.
}
\description{
Plots the triangular fuzzy numbers
defining the Likert-scale dictionary.
}
