% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prototypes.R
\name{plot_prototypes}
\alias{plot_prototypes}
\title{Plot Cluster Prototypes}
\usage{
plot_prototypes(
  object,
  cluster = 1,
  use_var_names = FALSE,
  var_names = NULL,
  ...
)
}
\arguments{
\item{object}{An object of class "fcmTFN".}

\item{cluster}{Integer cluster to plot.}

\item{use_var_names}{Logical.}

\item{var_names}{Optional variable names.}

\item{...}{Additional graphical parameters.}
}
\value{
Invisibly returns NULL.
}
\description{
Visualizes cluster prototypes as interval plots
using triangular fuzzy numbers (l, c, r).
}
