% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xb.R
\name{plot_xb}
\alias{plot_xb}
\title{Plot Xie-Beni Index}
\usage{
plot_xb(object, mark_optimal = TRUE, type = "b", ...)
}
\arguments{
\item{object}{An object of class "fcmTFN".}

\item{mark_optimal}{Logical. Whether to highlight the optimal k.}

\item{type}{Plot type (default = "b").}

\item{...}{Additional graphical parameters.}
}
\value{
Invisibly returns NULL.
}
\description{
Plots the Xie-Beni validity index
across candidate numbers of clusters.
}
